backend/requirements.txt
Flask
assemblyai
google-generativeai
python-dotenv
